Reran ant genrecord and then tidied up a fuckton of manual ColumnExp creation and usage. The new type-safe Where() clause caught three niggling bits. Two were passing a byte where the column type was a ByteEnum, which works, but is no longer supported. (Passing an actual enum in that location was only recently made to work, so passing the byte directly was necessary naughtiness.) The third niggling bit was someone passing an Integer where a byte was needed. I presume the JDBC driver just sorted that out, but it's best not to rely on that kind of lenience.

+++ src/java/com/threerings/msoy/item/server/ItemServlet.java
@@ -287,7 +287,7 @@ public class ItemServlet extends MsoyServiceServlet
         throws ServiceException
     {
         requireSupportUser();
-        _itemFlagRepo.removeItemFlags(iitem.type.toByte(), iitem.itemId);
+        _itemFlagRepo.removeItemFlags(iitem.type, iitem.itemId);
     }
 
     // from ItemService interface
